# infra/contracts/s9-openapi.yaml
openapi: 3.1.0
info:
  title: worldview-gateway
  version: 2025.6.0
paths:
...
  /v1/threads/{thread_id}:
    get:
      summary: Get Thread
      description: Get a single conversation thread (proxy to S8).
      operationId: get_thread_v1_threads__thread_id__get
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Thread V1 Threads  Thread Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Thread
      description: Delete a conversation thread (proxy to S8).
      operationId: delete_thread_v1_threads__thread_id__delete
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Delete Thread V1 Threads  Thread Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Thread
      description: 'Proxy PATCH /v1/threads/{thread_id} → S8 PATCH /api/v1/threads/{thread_id}.


        PLAN-0051 Wave E / T-E-5-06.


        Used by the chat UI to rename a thread inline (double-click on the

        sidebar title). Body is forwarded unchanged so future patchable fields

        don''t require a gateway change.'
      operationId: update_thread_v1_threads__thread_id__patch
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Update Thread V1 Threads  Thread Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...